requirements.txt
pdfminer.six==20221105
numpy==1.25.2
scikit-learn==1.3.0